SMG: Fix Integer vs int bug
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@28742 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/join/SMGJoinFields.java
@@ -134,7 +134,7 @@ class SMGJoinFields {
       if (// if there is none (meaning the block got shortened from the start)
           newNullBlock == null ||
           // or the new block has different size (got shortened from the end)
-          newNullBlock != origEdge.getValue()) {
+          newNullBlock.intValue() != origEdge.getValue().intValue()) {
 
         // check whether the block really got smaller and not bigger on its end
         // there is also un-checked possibility that the block got bigger on its start
